backend/requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
pandas>=2.2,<3.0
numpy>=1.26,<3.0
scikit-learn>=1.5,<2.0
